pyproject.toml
[project]
name = "MADSci_Examples"
version = "0.1.0"
description = "Examples for the Modular Autonomous Discovery for Science (MADSci) project."
authors = [
    {name = "Ryan Lewis", email = "ryan.lewis@anl.gov"},
]
dependencies = ["madsci.client>=0.4.0,<0.5.0", "madsci.common>=0.4.0,<0.5.0", "madsci.node_module>=0.4.0,<0.5.0", "ipykernel>=6.29.5", "httpx>=0.28.1"]
requires-python = ">=3.9.1"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
